Refactor class names (Transformer and Verifier)
