+++ src/main/java/com/github/dakusui/pcond/core/fluent/verifiers/NumberMatcherBuilder.java
@@ -1,0 +1,4 @@
+package com.github.dakusui.pcond.core.fluent.verifiers;
+
+public class NumberMatcherBuilder {
+}
