--- src/main/java/com/github/dakusui/pcond/core/matchers/chckers/NumberMatcherBuilder.java
@@ -1,4 +1,0 @@
-package com.github.dakusui.pcond.core.matchers.chckers;
-
-public class NumberMatcherBuilder {
-}
